refactor(test): split base_messages test files under cap
Tier 3 of #529: split three oversized test files in
crates/services/src/cell/service/base_messages under the 700-line
hard cap (CLAUDE.md "File organization"). Pure behavior-preserving
test moves — no test logic or assertions changed, #[test] counts
preserved (general 18, bandolier_sync 8, player_init 13).

general.rs (1330) → grouped by message-family theme:
  - movement.rs        position/teleport validation (6)
  - item_events.rs     item move/equip/use content events (5)
  - minigame.rs        minigame result chains (2)
  - trade_disconnect.rs disconnect/destroy trade cancel (2)
  - general.rs (kept)  destroy-flush, advance-ring, init name (3)

bandolier_sync.rs (747) → grouped by sync concern:
  - bandolier_sync.rs (kept)     equip/unequip animation (3)
  - bandolier_sync_ammo_type.rs  AmmoTypeId emit (3)
  - bandolier_sync_reload.rs     reload-on-activate (2)

player_init/tests.rs (900) → tests/ directory, one file per
nested theme module (super::super::* depth unchanged):
  - state_field_restore.rs         (3)
  - relog_mission_resurrection.rs  (1)
  - system_options_assignment.rs   hydrate/assignment guards (5)
  - reload_on_activate.rs          login reload gate guards (4)
