+++ crates/services/src/cell/service/base_messages/tests/mod.rs
@@ -4,7 +4,13 @@ use cimmeria_entity::cell_entity::BandolierItem;
 
 mod ability_granted_trainer_resend;
 mod bandolier_sync;
+mod bandolier_sync_ammo_type;
+mod bandolier_sync_reload;
 mod bandolier_update;
 mod general;
 mod gm_spawn_ready;
+mod item_events;
+mod minigame;
+mod movement;
 mod request_entity_update;
+mod trade_disconnect;
